feat(ui): allow fixing/removing mistagged faces in videos
Extract AssignFaceDialog into a shared, generalized component: it now
operates on a set of face IDs, auto-detects the video representative-frame
thumbnail vs. the photo bounding-box crop for its preview, and offers a
permanent-delete action via PurgeFacesDialog. Wire it into VideoFacePanel
through a per-row edit action (targeting all of a person's detected faces in
the video) and keep the video marker strip in sync via onFacesChanged.

Relates to #119
